show relations diagram in spark page
