+++ src/plugins/spark/components/Dialog/RelationsDialog.vue
@@ -1,49 +1,19 @@
 <script lang="ts">
-import { select as d3Select } from 'd3-selection';
-import { graphlib, render as dagreRender } from 'dagre-d3';
-import { saveSvgAsPng } from 'save-svg-as-png';
-import { setTimeout } from 'timers';
-import Vue from 'vue';
-import { Component, Prop, Ref } from 'vue-property-decorator';
-import { Watch } from 'vue-property-decorator';
+import { Component, Prop } from 'vue-property-decorator';
 
 import DialogBase from '@/components/Dialog/DialogBase';
-import { showBlockDialog } from '@/helpers/dialog';
-import { sparkStore } from '@/plugins/spark/store';
 import { RelationEdge, RelationNode } from '@/plugins/spark/types';
 
 
-const LABEL_HEIGHT = 50;
-const LABEL_WIDTH = 150;
-const INVERTED = [
-  'input',
-  'reference',
-  'sensor',
-];
-
 @Component
 export default class RelationsDialog extends DialogBase {
-  exportBusy = false;
-  graphObj: any = null;
-  availableHeight = 0;
-  availableWidth = 0;
-
-  @Ref()
-  readonly svg!: SVGGraphicsElement;
-
-  @Ref()
-  readonly diagram!: SVGGraphicsElement;
-
-  @Ref()
-  readonly toolbar!: Vue;
-
   @Prop({ type: String, required: true })
   readonly serviceId!: string;
 
-  @Prop({ type: Array, default: [] })
+  @Prop({ type: Array, required: true })
   readonly nodes!: RelationNode[];
 
-  @Prop({ type: Array, default: [] })
+  @Prop({ type: Array, required: true })
   readonly edges!: RelationEdge[];
 
   @Prop({ type: String, default: 'Block Relations' })
@@ -52,132 +22,11 @@ export default class RelationsDialog extends DialogBase {
   @Prop({ type: Boolean, default: false })
   public readonly hideUnrelated!: boolean;
 
-  get drawnNodes(): RelationNode[] {
-    return [...new Set(this.edges.flatMap(edge => [edge.target, edge.source]))]
-      .map(id => this.nodes.find(node => node.id === id) || { id, type: '???' });
-  }
-
-  get loneNodes(): RelationNode[] {
-    return this.nodes.filter(node => !this.drawnNodes.find(n => n.id === node.id));
-  }
-
-  @Watch('edges', { immediate: true })
-  display(newV: RelationEdge[], oldV: RelationEdge[] | null): void {
-    if (newV && JSON.stringify(newV) !== JSON.stringify(oldV)) {
-      setTimeout(() => this.calc() && setTimeout(() => this.draw(), 100), 100);
-    }
-  }
-
-  calc(): boolean {
-    const nodeTemplate = (id: string, type: string): string => {
-      return `
-        <div style="width: ${LABEL_WIDTH}px; height: ${LABEL_HEIGHT}px" ">
-          <div class="type">${type}</div>
-          <div class="id">${id}</div>
-        </div>
-        `;
+  get contentStyle(): Mapped<string> {
+    return {
+      height: `${window.innerHeight - 50}px`,
+      width: `${window.innerWidth}px`,
     };
-
-    const obj = new graphlib
-      .Graph({ multigraph: true, compound: true })
-      .setGraph({ marginx: 20, marginy: 20 });
-
-    const nodes = this.hideUnrelated
-      ? this.drawnNodes
-      : [...this.loneNodes, ...this.drawnNodes];
-
-    nodes.forEach(node => {
-      obj.setNode(node.id, {
-        id: node.id,
-        label: nodeTemplate(node.id, node.type),
-        labelType: 'html',
-        width: LABEL_WIDTH,
-        height: LABEL_HEIGHT,
-        rx: 5,
-        ry: 5,
-      });
-    });
-
-    this.edges.forEach(edge => {
-      const label = edge.relation[0].replace(/Id$/, '');
-      const [source, target] = INVERTED.includes(label)
-        ? [edge.target, edge.source]
-        : [edge.source, edge.target];
-
-      obj.setEdge(source, target, {
-        label,
-        labelStyle: 'fill: white; stroke: none;',
-        style: 'fill: none; stroke: red; stroke-width: 1.5px;',
-        arrowheadStyle: 'fill: red; stroke: red;',
-      },
-        edge.relation[0]);
-    });
-
-    if (!this.hideUnrelated) {
-      const invisible = 'fill: transparent; stroke: none;';
-      const stackHeight = 6;
-
-      this.loneNodes.forEach((node, idx) => {
-        if (idx % stackHeight === 0) { return; }
-        obj.setEdge(node.id, this.loneNodes[idx - 1].id, {
-          label: '',
-          labelStyle: invisible,
-          style: invisible,
-          arrowheadStyle: invisible,
-        });
-      });
-    }
-
-    this.graphObj = obj;
-    return true;
-  }
-
-  draw(): void {
-    const renderFunc = new dagreRender();
-    try {
-      renderFunc(d3Select(this.diagram), this.graphObj);
-    } catch (e) {
-      // Workaround for a bug in FireFox where getScreenCTM() returns null for hidden or 0x0 elements
-      if (e.name === 'TypeError') {
-        renderFunc(d3Select(this.diagram), this.graphObj);
-      } else {
-        throw e;
-      }
-    }
-    const outGraph = this.graphObj.graph();
-    const toolbarHeight = this.toolbar.$el.clientHeight || 50;
-    this.availableHeight = window.innerHeight - toolbarHeight;
-    this.availableWidth = window.innerWidth;
-    this.svg.setAttribute('style', `min-width: ${outGraph.width}px;
-                                          min-height: ${outGraph.height};`);
-    this.svg.setAttribute('height', outGraph.height);
-    this.svg.setAttribute('width', outGraph.width);
-
-    this.$nextTick(() => {
-      // Here be dragons
-      // Dagre incorrectly renders the injected HTML as a "foreignObject" with 0x0 size
-      // The hacky, but working solution is to override the SVG properties
-      this.svg.querySelectorAll('foreignObject')
-        .forEach(el => {
-          const id = el.children[0].children[0].children[1].innerHTML;
-          el.addEventListener('click', () => this.openSettings(id));
-          el.setAttribute('width', `${LABEL_WIDTH}`);
-          el.setAttribute('height', `${LABEL_HEIGHT}`);
-          (el.parentElement as HTMLElement)
-            .setAttribute('transform', `translate(-${LABEL_WIDTH / 2}, -${LABEL_HEIGHT / 2})`);
-        });
-    });
-  }
-
-  exportDiagram(): void {
-    this.exportBusy = true;
-    // uses quasar "dark" as background color
-    saveSvgAsPng(this.svg, 'block-relations.png', { backgroundColor: '#282c34' })
-      .finally(() => this.exportBusy = false);
-  }
-
-  openSettings(id: string): void {
-    showBlockDialog(sparkStore.blocks(this.serviceId)[id]);
   }
 }
 </script>
@@ -185,67 +34,13 @@ export default class RelationsDialog extends DialogBase {
 <template>
   <q-dialog ref="dialog" maximized no-backdrop-dismiss @hide="onDialogHide">
     <q-card dark class="maximized bg-dark-bright">
-      <DialogToolbar ref="toolbar" @close="onDialogHide">
+      <DialogToolbar @close="onDialogHide">
         {{ title }}
-        <template #buttons>
-          <!-- Exporting is bugged right now. See https://github.com/BrewBlox/brewblox-ui/issues/638 -->
-          <q-btn
-            v-if="false"
-            :loading="exportBusy"
-            flat
-            rounded
-            label="export"
-            @click="exportDiagram"
-          />
-        </template>
       </DialogToolbar>
-
-      <div
-        :style="`
-          overflow: auto;
-          width: ${availableWidth}px;
-          height: ${availableHeight}px;
-          `"
-        class="row"
-      >
-        <q-space />
-        <svg ref="svg" class="col-auto">
-          <g ref="diagram" />
-        </svg>
-        <q-space />
-      </div>
+      <RelationsDiagram
+        :content-style="contentStyle"
+        v-bind="$props"
+      />
     </q-card>
   </q-dialog>
 </template>
-
-<style lang="stylus" scoped>
-/deep/ .node .id {
-  font-weight: 300;
-  font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
-  font-size: 14px;
-  color: black;
-  padding: 10px;
-  width: 100%;
-  text-align: center;
-  cursor: pointer;
-}
-
-/deep/ .node .type {
-  font-weight: 300;
-  font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
-  font-size: 12px;
-  color: green;
-  width: 100%;
-  text-align: center;
-  cursor: pointer;
-}
-
-/deep/ .node rect {
-  fill: #fff;
-  cursor: pointer;
-}
-
-/deep/ .node:hover rect {
-  fill-opacity: 0.8;
-}
-</style>
